Add ckks mul and sq.

+++ src/kernelprovider.cuh
@@ -21,7 +21,7 @@ namespace troy {
         }
 
         template <typename T>
-        static T* malloc(int length) {
+        static T* malloc(size_t length) {
             checkInitialized();
             T* ret;
             auto status = cudaMalloc((void**)&ret, length * sizeof(T));
@@ -37,7 +37,7 @@ namespace troy {
         }
 
         template <typename T>
-        static void copy(T* deviceDestPtr, const T* hostFromPtr, int length) {
+        static void copy(T* deviceDestPtr, const T* hostFromPtr, size_t length) {
             checkInitialized();
             auto status = cudaMemcpy(deviceDestPtr, hostFromPtr, length * sizeof(T), cudaMemcpyHostToDevice);
             if (status != cudaSuccess) 
@@ -45,7 +45,7 @@ namespace troy {
         }
 
         template <typename T>
-        static void copyOnDevice(T* deviceDestPtr, const T* deviceFromPtr, int length) {
+        static void copyOnDevice(T* deviceDestPtr, const T* deviceFromPtr, size_t length) {
             checkInitialized();
             auto status = cudaMemcpy(deviceDestPtr, deviceFromPtr, length * sizeof(T), cudaMemcpyDeviceToDevice);
             if (status != cudaSuccess) 
@@ -53,7 +53,7 @@ namespace troy {
         }
 
         template <typename T>
-        static void retrieve(T* hostDestPtr, T* deviceFromPtr, int length) {
+        static void retrieve(T* hostDestPtr, T* deviceFromPtr, size_t length) {
             checkInitialized();
             auto status = cudaMemcpy(hostDestPtr, deviceFromPtr, length * sizeof(T), cudaMemcpyDeviceToHost);
             if (status != cudaSuccess) 
@@ -61,7 +61,7 @@ namespace troy {
         }
 
         template <typename T>
-        static void memsetZero(T* devicePtr, int length) {
+        static void memsetZero(T* devicePtr, size_t length) {
             cudaMemset(devicePtr, 0, sizeof(T) * length);
         }
